refactor(upskill): extract upskill and skill commands into separate modu
Agent-Id: agent-584ff711-5103-495f-852c-2cf30020f034
Linked-Note-Id: a4471e48-2375-48c1-afd1-62706736e75e
